Change hotkey for 'Select All Occurrences' to Alt+F3
Ctrl+Shift+I is in conflict with the Inspect command during code
debugging, even if when the context is changed from 'In Windows' to
'Text Editing'.

Therefore, it is better to select new hotkey. Alt+F3 is like in Sublime
Text and it is absolutely free (checked with an Eclipse product with a
large list of plugins).