Change hotkey for 'Select Next Occurrence' to Ctrl+Alt+D
Ctrl+Shift+D is in conflict with the 'Match Pair Inward' command of the
Emmet plugin. Emmet is a popular plugin among web developers and when
installed the 'Select Next Occurence' command does not work.

Therefore, it is better to select new hotkey. Ctrl+Alt+D is still not
very different from Ctrl+Shift+D, it's free and there are also lots of
other Ctrl+Alt+<key> shortcuts for text editing in the Eclipse Platform
and AnyEdit plugin. So, it is a good group to be inside.